// ui/src/api/generated/models/usage-call-detail.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Halo
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 2.26.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { UsageCallItem } from './usage-call-item';
// May contain unused imports in some cases
// @ts-ignore
import type { UsageExecutionRecord } from './usage-execution-record';

/**
 * 
 * @export
 * @interface UsageCallDetail
 */
export interface UsageCallDetail {
    /**
     * 
     * @type {UsageCallItem}
     * @memberof UsageCallDetail
     */
    'call'?: UsageCallItem;
    /**
     * 
     * @type {Array<UsageExecutionRecord>}
     * @memberof UsageCallDetail
     */
    'executions'?: Array<UsageExecutionRecord>;
}